feat: upgrade pnpm to 10.26.0 and implement version validation script with tests
