+++ .github/workflows/release-sdk.yml
@@ -34,7 +34,7 @@ jobs:
       - name: Setup pnpm
         uses: pnpm/action-setup@v5
         with:
-          version: 10
+          version: 10.26.0
 
       - name: Install workspace dependencies
         run: pnpm install --frozen-lockfile
